Face_Recog_WEB/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.5
docopt==0.6.2
face-recognition-models==0.3.0
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.1
pipreqs==0.4.11
protobuf==3.19.4
requests==2.26.0
responses==0.21.0
torch
torchvision
zipp==3.8.0
pandas 
opencv-python ==4.6.0.66
matplotlib
tqdm
pyyaml
seaborn
thop
